bot/requirements.txt
python-telegram-bot[webhooks]==21.11
aiosqlite>=0.20.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.8.0
httpx>=0.28.0
anthropic>=0.49.0
openai>=1.60.0
trafilatura>=1.12.0
readability-lxml>=0.8.1
lxml>=5.0.0
lxml_html_clean>=0.2.0
apscheduler>=3.10.0
tomli_w>=1.0.0
